refactor: split schema.ts and monitoring.ts into modular sub-directories
- schema.ts (1083 lines, 12 tools) -> schema/objects.ts + schema/views.ts
- monitoring.ts (941 lines, 11 tools) -> monitoring/basic.ts + monitoring/analysis.ts
- introspection.ts (2234 lines, 12 tools) -> introspection/graph.ts + analysis.ts + migration.ts
- text.ts (1235 lines, 13 tools) -> text/search.ts + text/matching.ts
- Each group has a barrel index.ts re-exporting the public API
- Updated PostgresAdapter.ts import paths
- No functional changes
